+++ Emulator/Memory/MemoryTypes.h
@@ -222,13 +222,13 @@ MemoryConfig;
 
 typedef struct
 {
-    struct { long raw; double accumulated; } chipReads;
-    struct { long raw; double accumulated; } chipWrites;
-    struct { long raw; double accumulated; } slowReads;
-    struct { long raw; double accumulated; } slowWrites;
-    struct { long raw; double accumulated; } fastReads;
-    struct { long raw; double accumulated; } fastWrites;
-    struct { long raw; double accumulated; } kickReads;
-    struct { long raw; double accumulated; } kickWrites;
+    struct { isize raw; double accumulated; } chipReads;
+    struct { isize raw; double accumulated; } chipWrites;
+    struct { isize raw; double accumulated; } slowReads;
+    struct { isize raw; double accumulated; } slowWrites;
+    struct { isize raw; double accumulated; } fastReads;
+    struct { isize raw; double accumulated; } fastWrites;
+    struct { isize raw; double accumulated; } kickReads;
+    struct { isize raw; double accumulated; } kickWrites;
 }
 MemoryStats;
